requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
beartype==0.11.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
Django==4.1.2
exceptiongroup==1.0.4
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
multidict==6.0.2
packaging==22.0
paramiko==2.12.0
pluggy==1.0.0
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.21
PyNaCl==1.5.0
pytest==7.2.0
pytest-mock==3.10.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.4
urllib3==1.26.12
waitress==2.1.2
whitenoise==6.2.0
yarl==1.8.1
